requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pandas==2.2.2
numpy==1.26.4
plotly==5.23.0
streamlit==1.37.0
python-multipart==0.0.9
pydantic==2.8.2
openai==1.37.0
python-pptx==0.6.23
kaleido==0.2.1
python-dotenv==1.0.1
